+++ src/main/java/com/springhrms/model/Region.java
@@ -15,14 +15,14 @@ import javax.persistence.Table;
 @Table(name = "regions")
 public class Region {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	@Column(name = "region_id")
-	private int regionId;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "region_id")
+    private long regionId;
 
-	@Column(name = "region_name")
-	private String regionName;
+    @Column(name = "region_name")
+    private String regionName;
 
-	@OneToMany(mappedBy = "region", cascade = CascadeType.ALL)
-	private Set<Country> countries;
+    @OneToMany(mappedBy = "region", cascade = CascadeType.ALL)
+    private Set<Country> countries;
 }
